Refactor smtp specific email handling
